requirements.txt
streamlit
pymupdf
sentence-transformers
scikit-learn
faiss-cpu
python-dotenv
google-generativeai
